+++ src/App.cpp
@@ -121,11 +121,11 @@ bool App::IsPressed(string k){
 }
 
 void App::PhysicsUpdate(){
-  if(1000/pFPS <= SDL_GetTicks()-pTime){
+  if((unsigned)(1000/pFPS) <= SDL_GetTicks()-pTime){
     for(int i = 0; i < particleSystems.size();){
       particleSystems[i]->Update(this);
 
-      if(SDL_GetTicks() >= particleSystems[i]->GetTime()){
+      if(SDL_GetTicks() >= (unsigned)particleSystems[i]->GetTime()){
         particleSystems.erase(particleSystems.begin()+i);
       }
       else{
